+++ crates/client/sync/src/l1.rs
@@ -1,6 +1,6 @@
 //! Contains the necessaries to perform an L1 verification of the state
 
-use std::sync::{Arc, RwLock};
+use std::sync::Arc;
 
 use anyhow::{bail, Context, Result};
 use ethers::contract::{abigen, EthEvent};
@@ -9,7 +9,7 @@ use ethers::types::transaction::eip2718::TypedTransaction;
 use ethers::types::{Address, BlockNumber as EthBlockNumber, Filter, TransactionRequest, I256, U256, U64};
 use ethers::utils::hex::decode;
 use futures::stream::StreamExt;
-use lazy_static::lazy_static;
+use mc_db::{DeoxysBackend, WriteBatchWithTransaction};
 use mp_felt::Felt252Wrapper;
 use primitive_types::H256;
 use prometheus_endpoint::prometheus::core::Number;
@@ -19,17 +19,9 @@ use serde_json::Value;
 use starknet_api::hash::StarkHash;
 
 use crate::metrics::block_metrics::BlockMetrics;
-use crate::utility::{convert_log_state_update, l1_core_address};
+use crate::utility::convert_log_state_update;
+use crate::utils::channel_wait_or_graceful_shutdown;
 use crate::utils::constant::LOG_STATE_UPDTATE_TOPIC;
-
-lazy_static! {
-    /// Shared latest L2 state update verified on L1
-    pub static ref ETHEREUM_STATE_UPDATE: Arc<RwLock<L1StateUpdate>> = Arc::new(RwLock::new(L1StateUpdate {
-        block_number: u64::default(),
-        global_root: StarkHash::default(),
-        block_hash: StarkHash::default(),
-    }));
-}
 
 /// Contains the Starknet verified state on L1
 #[derive(Debug, Clone, Deserialize)]
@@ -52,14 +44,15 @@ pub struct LogStateUpdate {
 pub struct EthereumClient {
     provider: Arc<Provider<Http>>,
     url: Url,
+    l1_core_address: Address,
 }
 
 /// Implementation of the Ethereum client to interact with L1
 impl EthereumClient {
     /// Create a new EthereumClient instance with the given RPC URL
-    pub async fn new(url: Url) -> Result<Self> {
+    pub async fn new(url: Url, l1_core_address: Address) -> Result<Self> {
         let provider = Provider::<Http>::try_from(url.as_str())?;
-        Ok(Self { provider: Arc::new(provider), url })
+        Ok(Self { provider: Arc::new(provider), url, l1_core_address })
     }
 
     /// Get current RPC URL
@@ -82,7 +75,7 @@ impl EthereumClient {
     /// Get the block number of the last occurrence of a given event.
     pub async fn get_last_event_block_number(&self) -> anyhow::Result<u64> {
         let topic = H256::from_slice(&hex::decode(&LOG_STATE_UPDTATE_TOPIC[2..])?);
-        let address = l1_core_address();
+        let address = self.l1_core_address;
         let latest_block = self.get_latest_block_number().await.expect("Failed to retrieve latest block number");
 
         // Assuming an avg Block time of 15sec we check for a LogStateUpdate occurence in the last ~24h
@@ -105,7 +98,7 @@ impl EthereumClient {
     /// Get the last Starknet block number verified on L1
     pub async fn get_last_block_number(&self) -> anyhow::Result<u64> {
         let data = decode("35befa5d")?;
-        let to: Address = l1_core_address();
+        let to: Address = self.l1_core_address;
         let tx_request = TransactionRequest::new().to(to).data(data);
         let tx = TypedTransaction::Legacy(tx_request);
         let result = self.provider.call(&tx, None).await.expect("Failed to get last block number");
@@ -119,7 +112,7 @@ impl EthereumClient {
     /// Get the last Starknet state root verified on L1
     pub async fn get_last_state_root(&self) -> Result<StarkHash> {
         let data = decode("9588eca2")?;
-        let to: Address = l1_core_address();
+        let to: Address = self.l1_core_address;
         let tx_request = TransactionRequest::new().to(to).data(data);
         let tx = TypedTransaction::Legacy(tx_request);
         let result = self.provider.call(&tx, None).await.expect("Failed to get last state root");
@@ -129,7 +122,7 @@ impl EthereumClient {
     /// Get the last Starknet block hash verified on L1
     pub async fn get_last_block_hash(&self) -> Result<StarkHash> {
         let data = decode("0x382d83e3")?;
-        let to: Address = l1_core_address();
+        let to: Address = self.l1_core_address;
         let tx_request = TransactionRequest::new().to(to).data(data);
         let tx = TypedTransaction::Legacy(tx_request);
         let result = self.provider.call(&tx, None).await.expect("Failed to get last block hash");
@@ -153,7 +146,7 @@ impl EthereumClient {
         block_metrics: Option<BlockMetrics>,
     ) -> anyhow::Result<()> {
         let client = self.provider.clone();
-        let address: Address = l1_core_address();
+        let address: Address = self.l1_core_address;
         abigen!(
             StarknetCore,
             "crates/client/sync/src/utils/abis/starknet_core.json",
@@ -165,11 +158,11 @@ impl EthereumClient {
 
         let mut event_stream = event_filter.stream().await.context("initiatializing event stream")?;
 
-        while let Some(event_result) = event_stream.next().await {
+        while let Some(event_result) = channel_wait_or_graceful_shutdown(event_stream.next()).await {
             let log = event_result.context("listening for events")?;
             let format_event =
                 convert_log_state_update(log.clone()).context("formatting event into an L1StateUpdate")?;
-            update_l1(format_event, block_metrics.clone());
+            update_l1(format_event, block_metrics.clone())?;
         }
 
         Ok(())
@@ -177,7 +170,7 @@ impl EthereumClient {
 }
 
 /// Update the L1 state with the latest data
-pub fn update_l1(state_update: L1StateUpdate, block_metrics: Option<BlockMetrics>) {
+pub fn update_l1(state_update: L1StateUpdate, block_metrics: Option<BlockMetrics>) -> anyhow::Result<()> {
     log::info!(
         "🔄 Updated L1 head: Number: #{}, Hash: {}, Root: {}",
         state_update.block_number,
@@ -189,11 +182,13 @@ pub fn update_l1(state_update: L1StateUpdate, block_metrics: Option<BlockMetrics
         block_metrics.l1_block_number.set(state_update.block_number.into_f64());
     }
 
-    {
-        let last_state_update = ETHEREUM_STATE_UPDATE.clone();
-        let mut new_state_update = last_state_update.write().expect("poisoned lock");
-        *new_state_update = state_update.clone();
-    }
+    let mut tx = WriteBatchWithTransaction::default();
+    DeoxysBackend::mapping()
+        .write_last_confirmed_block(&mut tx, state_update.block_number)
+        .context("setting l1 last confirmed block number")?;
+    DeoxysBackend::expose_db().write(tx).context("writing pending block to db")?;
+    log::debug!("update_l1: wrote last confirmed block number");
+    Ok(())
 }
 
 // /// Verify the L1 state with the latest data
@@ -231,14 +226,24 @@ pub fn update_l1(state_update: L1StateUpdate, block_metrics: Option<BlockMetrics
 // }
 
 /// Syncronize with the L1 latest state updates
-pub async fn sync(l1_url: Url, block_metrics: Option<BlockMetrics>) -> anyhow::Result<()> {
-    let client = EthereumClient::new(l1_url).await.context("creating ethereum client")?;
+pub async fn sync(l1_url: Url, block_metrics: Option<BlockMetrics>, l1_core_address: Address) -> anyhow::Result<()> {
+    // Clear L1 confirmed block at startup
+    {
+        let mut tx = WriteBatchWithTransaction::default();
+        DeoxysBackend::mapping()
+            .write_no_last_confirmed_block(&mut tx)
+            .context("clearing l1 last confirmed block number")?;
+        DeoxysBackend::expose_db().write(tx).context("writing pending block to db")?;
+        log::debug!("update_l1: cleared confirmed block number");
+    }
+
+    let client = EthereumClient::new(l1_url, l1_core_address).await.context("creating ethereum client")?;
 
     log::info!("🚀 Subscribed to L1 state verification");
 
     // Get and store the latest verified state
     let initial_state = EthereumClient::get_initial_state(&client).await.context("getting initial ethereum state")?;
-    update_l1(initial_state, block_metrics.clone());
+    update_l1(initial_state, block_metrics.clone())?;
 
     // Listen to LogStateUpdate (0x77552641) update and send changes continusly
     let start_block =
@@ -276,9 +281,10 @@ mod l1_sync_tests {
     }
 
     #[tokio::test]
+    #[ignore]
     async fn test_starting_block() {
         let url = Url::parse(eth_rpc::MAINNET).expect("Failed to parse URL");
-        let client = EthereumClient::new(url).await.expect("Failed to create EthereumClient");
+        let client = EthereumClient::new(url, H160::zero()).await.expect("Failed to create EthereumClient");
 
         let start_block =
             EthereumClient::get_last_event_block_number(&client).await.expect("Failed to get last event block number");
@@ -286,9 +292,10 @@ mod l1_sync_tests {
     }
 
     #[tokio::test]
+    #[ignore]
     async fn test_initial_state() {
         let url = Url::parse(eth_rpc::MAINNET).expect("Failed to parse URL");
-        let client = EthereumClient::new(url).await.expect("Failed to create EthereumClient");
+        let client = EthereumClient::new(url, H160::zero()).await.expect("Failed to create EthereumClient");
 
         let initial_state = EthereumClient::get_initial_state(&client).await.expect("Failed to get initial state");
         assert!(!initial_state.global_root.0.is_empty(), "Global root should not be empty");
@@ -297,6 +304,7 @@ mod l1_sync_tests {
     }
 
     #[tokio::test]
+    #[ignore]
     async fn test_event_subscription() -> Result<(), Box<dyn std::error::Error>> {
         abigen!(
             IERC20,