+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Next release
 
 - refactor: enforce Felt type as much as possible into compute_hash.rs (changes in cascade)
+- fix(rpc): fixed block storage column
+- chore: update dependencies
+- fix(hashers): cleaned hashers using types core hashers and Felt
+- refactor: remove substrate block storage
 - feat(infra): Added boilerplate to deploy a grafana/prometheus dashboard
 - refacor: use db hash
 - refactor: l2-sync